requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
jinja2>=3.1.3
python-docx>=1.1.0
openpyxl>=3.1.2
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
aiofiles>=23.2.1
gunicorn>=21.2.0
anthropic>=0.39.0
python-dotenv>=1.0.0
